third_party/blink/web_tests/virtual/third-party-storage-partitioning/external/wpt/service-workers/service-worker/partitioned-matchAll.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] ServiceWorker's matchAll() is partitioned
  assert_equals: expected 1 but got 2
Harness: the test ran to completion.
